Introduce Router interface
Router interface allows library users to provide thier own router
implementations to be used - eg. predicate based router.
